Add Home view and router-view to display ConstellationDetail
